Make ABSTRACT CLASS to make code clean.
